\name{FRRFdata}
\alias{FRRFdata}
\title{
  FRRF data generated by the FastOcean FRRF (Fast Repetition Rate Fluorometer).
}

\usage{
data(FRRFdata)
}


\value{
  a data set with columns: 
\itemize{
  \item \code{file}, the name of the original data file
  \item \code{Saq}, a counter
  \item \code{E}, the photon flux density, [micromol photons / m2/s]
  \item \code{Start, s}, the start time of data point (actual and in seconds)
  \item \code{Chl}, the chlorophyll concentration, [mgChl/m3]
  \item \code{ADC}, ??
  \item \code{rP_measured, rP_fitted}, relative PSII electron transfer rate (relative whole chain electron transport rate)=E x Fq'/Fm' 
  \item \code{JPII}, PSII flux, [electrons PSII^{-1} s^{-1}]
  \item \code{JVPII}, PSII flux per unit volume [mmol e/ m3/ hour], = Fq/ Fm* a_LHII_0 * E * convJVPII
  \item \code{Fo}, (Fo') fluorescence at zeroth flashlet of an ST measurement when C = 0 (minimal fluorescence under ambient light) [-]
  \item \code{Fm}, (Fm') fluorescence at n-th flashlet of an ST measurement when C = 1 (under ambient light, maximal fluorescence) , [-]
  \item \code{Fq/Fm}, (Fm'-F')/Fm'; fluorescence parameter providing an estimate of PSII efficiency under ambient light, [-]
  \item \code{C}, (Cp, Cj or Cl) fraction of RCII that are in the closed state (estimated as 1-qP, 1-qJ or 1-qL) [-]
  \item \code{p}, connectivity parameter
  \item \code{Rsigma}, probability of an RCII being closed during the first flashlet of an ST saturation phase, [-]
  \item \code{Sigma}, absorption cross section of PSII photochemistry (under ambient light) [nm^2/PSII]
  \item \code{CSQ} ???
  \item \code{TauES} timescale for reopening of a closed RCII with an empty QB site [microsec??]
  \item \code{NPQ} nonphotochemical quenching 
  \item \code{NSV}, Normalized Stern-Volmer (SV), alternative to NPQ which treats sum of NP processes as SV quenchers of PSII photochemistry
  \item \code{QR, Qo, Qm, QoSE, QmSE, QSE, QSE.ratio, Qo.points, Qo,slope, Qo.intercept, Qm.points, Qm.slope, Qm.intercept} ??
  \item \code{Fblanc} background fluorescence (filtered water)
  \item \code{Fo_uc, Fm_uc} uncorrected values of Fo, Fm, as in original file>
  \item \code{JVPII_uc} uncorrected values of JVPII.[mol electrons (PSII m-3) d-1] 
  \item \code{Fq} = Fm - Fo
  \item {a_LHII}, the cross-sectional surface of the PSII system, = Fo*Fm/Fq*(Ka/10^6)
  \item {FqFm} = Fq/Fm
}
}

\description{
 Dataset obtained during the Footprint cruise (64PE504) with the Dutch Research Vessel Pelagia.  
 
 Coordinates of the sample were 51.7892dgN, 3.0877dgE; 36 m water depth; the sample was taken at  2022-06-28 13:25:49. 
 
 The FRRF measurements were performed based on surface water taken from niskin bottles, taken together with CTD casts. 
 
 The apparatus used was the FastOcean from Chelsea Technologies Ltd (FastACT2). The light source used was "LED combination A (450 nm alone)".  
 
 The datafile FRRFdata.raw is the raw data file that has been read by function \link{readFRRF}; 
 FRRFdata has been processed with function \link{standardizeFRRF}.
}

\examples{

# this is how the data sets were created
\dontrun{
 fpath        <- system.file("extdata", package="FRRF")
 FRRFdata_raw <- readFRRF(file="20220628-150602.csv", dir=fpath)
 FRRFdata     <- standardizeFRRF(FRRFdata_raw, 
                                 Fblanc = 0.1)
}

head(FRRFdata)
FRRFdata     <-  standardizeFRRF(FRRFdata, 
                                 Fblanc = 0.1)
 
# fit one set to data...
 set  <- subset(FRRFdata, 
                subset = ( !is.na(p)))
 
 plot(set$E, set$JVPII, 
      xlab = "light intensity, uEinst/m2/s", 
      ylab = "photosynthesis, mmol phot/m3/hr", 
      main = " PI data", 
      pch = 18, cex = 2, las = 1, cex.lab = 1.2)
 
# the relationship between these variables
 pairs(set [,c("E", "Chl", "JVPII", "Sigma", "TauES", "Fo", "Fm", "Fq/Fm")])

# fit the Webb model and PG model
 FitWebb <- fitPI("Webb", set$E, set$JVPII)
 FitPG   <- fitPI("PG",   set$E, set$JVPII)
 FitJP   <- fitPI("JP",   set$E, set$JVPII)
 FitEP   <- fitPI("EP",   set$E, set$JVPII)

 FitWebb.I <- fitPI("Webb", set$E, set$JVPII, normalized = TRUE)
 FitPG.I   <- fitPI("PG",   set$E, set$JVPII, normalized = TRUE)
 FitJP.I   <- fitPI("JP",   set$E, set$JVPII, normalized = TRUE)
 FitEP.I   <- fitPI("EP",   set$E, set$JVPII, normalized = TRUE)

# make predictions 

 PARpred  <- seq(0, 600, length.out = 100)
 predWebb <- predict(FitWebb, I = PARpred)
 predPG   <- predict(FitPG,   I = PARpred)
 predJP   <- predict(FitJP,   I = PARpred)
 predEP   <- predict(FitEP,   I = PARpred)
 
 predWebb.I <- predict(FitWebb.I, I = PARpred, normalized = TRUE)
 predPG.I   <- predict(FitPG.I,   I = PARpred, normalized = TRUE)
 predJP.I   <- predict(FitJP.I,   I = PARpred, normalized = TRUE)
 predEP.I   <- predict(FitEP.I,   I = PARpred, normalized = TRUE)

# plot predictions along with the data 
# not normalized 

 plot(set$E, set$JVPII, 
      xlab = "light intensity, uEinst/m2/s", 
      ylab = "photosynthesis, mmol phot/m3/hr", 
      main = "unnormalized fit", 
      pch = 18, cex = 2, las = 1, cex.lab = 1.2)
      
 lines(PARpred, predWebb) 
 lines(PARpred, predPG, col = "red") 
 lines(PARpred, predJP, col = "blue") 
 lines(PARpred, predEP, col = "green") 
 
 legend("bottomright", 
        legend = c("Webb",   "PG",   "JP",    "EP"), 
        col    = c("black", "red", "blue", "green"), lty = 1)

# normalized
 plot(set$E, set$JVPII, 
      xlab = "light intensity, uEinst/m2/s", 
      ylab = "photosynthesis, mmol phot/m3/hr", 
      main = "normalized fit", 
      pch = 18, cex = 2, las = 1, cex.lab = 1.2)
      
 lines(PARpred, predWebb.I) 
 lines(PARpred, predPG.I, col = "red") 
 lines(PARpred, predJP.I, col = "blue") 
 lines(PARpred, predEP.I, col = "green") 
 
 legend("bottomleft", 
        legend = c("Webb",   "PG",   "JP",   "EP"), 
        col    = c("black", "red", "blue", "green"), lty = 1)
}

\seealso{
  \code{\link{fitPI}} for fitting.
}
\author{
  Karline Soetaert <karline.soetaert@nioz.nl>
}
\keyword{utilities}

