\name{readFRRF}
\alias{readFRRF}

\title{
  Reads output from the Actnow or Labstaf apparatus.
}

\description{
  \code{readFRRF} reads text data generated by the FastOcean FRRF (Fast Repetition Rate Fluorometer) or Labstaf (Single turnover active fluorimeter). 
}

\usage{
readFRRF (file, dir="", specs=NULL, attr=NULL, txt="csv",  ...)
}

\arguments{
  \item{file }{The name of a the file to read. This name will be kept as a column of the output. 
  }
  \item{dir }{The name of the directory holding the file. Will not be kept in the output.
  }
  \item{specs }{A list with any extra information to be added as an extra column to the output (e.g. station names).
  }
  \item{attr }{A list with information to be added to the attributes of the output.
  }
  \item{txt }{How the data have to be read. This refers to the table read function of R, i.e. "csv" refers to \link{read.csv}, "csv2" refers to \link{read.csv2}, "table" refers to \link{read.table}, "delim" refers to \link{read.delim}, "delim2" refers to \link{read.delim2}
  }
  \item{... }{Any other arguments passed to the reading function , i.e. dec, quote, fill, etc.
  }
}
\note{
Two example files are in the package extdata directory. They are: 49O_40m_FRRF.csv (generated with FRRF) and 49O_40m_LS.txt (Labstaf). The package directory can be grasped with function system.file. See examples

}
\value{
  Returns a data.frame, with the first column the name of the data file. Other columns depend on the type of input file, e.g. 
  
  for the FRRF this is:
  
  "Saq", "E", "Start", "s", "Chl", "ADC", "rP_measured", 
  "rP_fitted",  "JPII", "JVPII",  "Fo", "Fm", "Fv/Fm", "C", "p"  ,         
 "RSigma",  "Sigma", "CSQ", "TauES",  "NPQ", "NSV", "X.1",  "QR" ,         
 "Qo", "Qm",  "QoSE", "QmSE", "QSE",  "QSE.ratio",  "X.2",  "Qo.points",   
 "Qo.slope", "Qo.intercept", "Qm.points", "Qm.slope", "Qm.intercept", "date" 
 
 for the Labstaf, this is:
 
"file", "E", "Fb", "rP", "rP.fit", "JVPII", "GOPII", "JPII", "Fo" ,        
"Fm",  "Fq.", "Fq..Fm.", "Fq..Fmc.",  "Fv..Fmc.", "Fq..Fv.", 
"EkS" , "Ekt.", "Ekt", "AlphaPII.",  "SigmaPII.",   "Rho.",  
"TauS.", "Taut.", "Fo", "Fm", "Fv", "Fv.Fm",      
"Fv.Fmc", "Fv..Fmc..1",  "EkS.1", "AlphaPII", "SigmaPII", "Rho", "TauS", 
"rP.1",  "JVPII.1", "GOPII.1", "JPII.1",  "F..1", "Fm..1", "Fq..1", 
"Fq..Fm..1", "Fq..Fmc..1",  "Fv..Fmc..2",  "Fq..Fv..1", "EkS.2", "Ekt..1", 
"Ekt.1", "AlphaPII..1" "SigmaPII..1","Rho..1", "TauS..1", "Taut..1", "date" 

Other columns may be added, depending on the settings of argument "specs".


The attributes of the object will contain: 
\itemize{
 \item system: the name of the apparatus (act2 or LabStaf), 
 \item file: the name of the file
 \item date: the mean date of creation
 \item settings: settings of the apparatus
 \item fitted pars: the (preliminary) fit parameters of the PI curve
 \item processing: text keeping track of processing of the data
 \item standardized: true or false whether data have been starndarized or not 
}

The attributes of the object also contain all information passed with argument "attr".
}

\examples{

# path of the external data of the package
fpath <- system.file("extdata", package="dtPP")

## =============================================================
## Read the FRRF file
## =============================================================

F1 <- readFRRF(file = "49O_40m_FRRF.csv", 
               dir  = fpath)
head(F1, n=2)

AF1 <- attributes(F1)
AF1$settings
AF1$fitted.pars

AF1$standardized  # raw FRRF data are not standardized by default

## =============================================================
## Read the labstaf file
## =============================================================

F2 <- readFRRF(file = "49O_1m_LS.txt",    
               dir  = fpath, 
               txt  = "delim")
head(F2, n=2)

attributes(F2)$standardized  # LabStaf data are already standardized

## =============================================================
## Read the files, add extra information to table and attributes
## =============================================================

F1 <- readFRRF(file  = "49O_40m_FRRF.csv", 
               dir   = fpath, 
               specs = list(station="49O", depth=40), 
               attr  = list(author = "Auria"))

head(F1, n=2)

AF1 <- attributes(F1)
AF1$settings
AF1$author

}

\seealso{
   \code{\link{standardizeFRRF}} for subsequent standardization and estimating JVPII
    
   \code{\link{fitPI}} for fitting.

  \code{\link{fWebb}} for the various PI functions.

  \code{\link{LABSTAFdata}} for a data set generated with the labSTAF.
  
  \code{\link{FRRFdata}} for fitting with normalized and not normalized data.
  
  \code{\link{modFit}} for options of fitting algorithm
}
\author{
  Karline Soetaert <karline.soetaert@nioz.nl>
}
\keyword{utilities}
