\name{LABSTAFdata}
\alias{LABSTAFdata}
\title{
  A set of P-I data generated by the labSTAF appparatus.
}

\usage{
data(LABSTAFdata)
}


\value{
  a data set with columns:  
  \itemize{
  \code{file}, the name of the original file with the data (unique identifier),
  \code{station}, the name of the station,
  \item{depth}, depth below the water surface, m
  \item{E}, the photon flux density, [micromol photons / m2/s]
  \item{Fb}, the base photosynthesis used
  \item{rP, rP.fit}, relative PSII electron transfer rate (relative whole chain electron transport rate)
  \item{JVPII},  PSII flux per unit volume, [mol electrons (PSII m-3) d-1]
  \item{GOPII}, 
  \item{JPII}, PSII flux, [electrons PSII^-1 s^-1]
  \item{Fo}, (Fo') fluorescence at zeroth flashlet of an ST measurement when C = 0 (minimal fluorescence under ambient light) [-]
  \item{Fm},  (Fm') fluorescence at n-th flashlet of an ST measurement when C = 1 (under ambient light, maximal fluorescence) , [-] 
  \item{Fq.}, (Fm`-F)
  \item{Fq..Fm.},  ((Fm'-F')/Fm'); fluorescence parameter providing an estimate of PSII efficiency under ambient light, [-]`
  \item{Fq..Fmc.}, corrected value
  \item{Fq..Fv.}, \item{Eks}, \item{Ekt}, 
  \item{AlphaPII.}, 
  \item{SigmaPII.}, absorption cross section of PSII photochemistry (under ambient light) [nm^2/PSII]
  \item{Rho.}, ..
}
}

\description{
 to be added.
}

\examples{
 head(LABSTAFdata)

# the relationship between some variables
 pairs(LABSTAFdata [,c("E", "rP", "JVPII", "Fo", "Fm", "Fq..Fm.", "Fq..Fmc.", 
                       "SigmaPII.", "Rho.")])
 
## =======================
## fit one set to data...
## =======================

 set  <- subset(LABSTAFdata, 
                subset = (file == "59B_10m.txt" & 
                          ! is.na(JVPII)))

# fit the PI models to the data
# Webb, 1974 
 FitWebb <- fitPI(model      = "Webb", 
                  I          = set$E, 
                  response   = set$JVPII, 
                  normalized = TRUE)

# Platt, Galegos and Harrison, 1980  
 FitPG   <- fitPI(model      = "PG",   
                  I          = set$E, 
                  response   = set$JVPII, 
                  normalized = TRUE)

# Jassby and Platt, 1976
 FitJP   <- fitPI(model      = "JP",   
                  I          = set$E, 
                  response   = set$JVPII, 
                  normalized = TRUE)

# Eilers and Peeters, 1988
 FitEP   <- fitPI(model      = "EP",   
                  I          = set$E, 
                  response   = set$JVPII, 
                  normalized = TRUE)

# make predictions 

 PARpred  <- seq(0, 800, length.out = 100)
 predWebb <- predict(FitWebb, I = PARpred)
 predPG   <- predict(FitPG,   I = PARpred)
 predJP   <- predict(FitJP,   I = PARpred)
 predEP   <- predict(FitEP,   I = PARpred)

# plot predictions along with the data 

 plot(set$E, set$JVPII, 
      log = "x", xlab = "PAR", ylab = "p")
 lines(PARpred, predWebb) 
 lines(PARpred, predPG, col = "red") 
 lines(PARpred, predJP, col = "blue") 
 lines(PARpred, predEP, col = "green") 
 
 legend("bottomright", 
        legend = c("Webb",   "PG",   "JP",   "EP"), 
         col   = c("black", "red", "blue", "green"), lty = 1)

## =======================
## fit all sets to data...
## =======================

# fit the Webb model 
 FitWebb <- fitmultiPI(model     = "Webb", 
                      id         = LABSTAFdata$file, 
                      I          = LABSTAFdata$E, 
                      response   = LABSTAFdata$JVPII, 
                      normalized = TRUE)
                   
 plot(FitWebb)

}
\seealso{
  \code{\link{fitPI}} for fitting.
  \code{\link{FRRFdata}} for one raw dataset generated with the FRRF 
}
\author{
  Karline Soetaert <karline.soetaert@nioz.nl>
}
\keyword{utilities}

